panel admin index + css

+++ Projet/src/Controller/PlaceController.php
@@ -10,18 +10,19 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-#[Security("is_granted('ROLE_ADMIN'))")]
+#[Security("is_granted('ROLE_ADMIN')")]
 #[Route('/place')]
 class PlaceController extends AbstractController
 {
     #[Route('/', name: 'app_place_index', methods: ['GET'])]
     public function index(PlaceRepository $placeRepository): Response
     {
+
         return $this->render('place/index.html.twig', [
             'places' => $placeRepository->findAll(),
         ]);
     }
-
+    #[Security("(is_granted('CUSTOMERS'))")]
     #[Route('/new', name: 'app_place_new', methods: ['GET', 'POST'])]
     public function new(Request $request, PlaceRepository $placeRepository): Response
     {
@@ -48,7 +49,7 @@ class PlaceController extends AbstractController
             'place' => $place,
         ]);
     }
-
+    #[Security("(is_granted('ROLE_ADMIN'))")]
     #[Route('/{id}/edit', name: 'app_place_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Place $place, PlaceRepository $placeRepository): Response
     {
@@ -66,14 +67,16 @@ class PlaceController extends AbstractController
             'form' => $form,
         ]);
     }
-
+    #[Security("(is_granted('ROLE_CUSTOMER'))")]
     #[Route('/{id}', name: 'app_place_delete', methods: ['POST'])]
     public function delete(Request $request, Place $place, PlaceRepository $placeRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$place->getId(), $request->request->get('_token'))) {
-            $placeRepository->remove($place, true);
+        if(!$place->isPayer()) {
+            if ($this->isCsrfTokenValid('delete' . $place->getId(), $request->request->get('_token'))) {
+                $placeRepository->remove($place, true);
+            }
         }
 
-        return $this->redirectToRoute('app_place_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_annonce_panier', [], Response::HTTP_SEE_OTHER);
     }
 }
